daemon: move claim `jsonrpc_` methods to an individual module
